feat: Safer epoll timeout
