+++ changelog/2202.changed.md
@@ -1,0 +1,1 @@
+`Epoll::wait` now takes `EpollTimeout` replacing `isize`.
